+++ src/main/java/com/sl/springboot/dao/Repository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sl.springboot.models.Book;
 
-public interface Repository extends CrudRepository<Book, Long>{
+public interface Repository extends CrudRepository<Book, Integer>{
